Add files via upload
the basic analysis of the data is done. data analysis like head, tail, describe and the shape of the dataset and and some other necessary are done. the null values are checked and  column analysis is also done. we took all the positive cases of monkey pox and found the common symptoms and showed in a pie chat and later changed the boolean values into int format.